add Flatten to tsutils pkg
